% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun_editor_input.R
\name{update_sun_editor_input}
\alias{update_sun_editor_input}
\title{Change the value of a SunEditor input on the client}
\usage{
update_sun_editor_input(
  session = shiny::getDefaultReactiveDomain(),
  input_id,
  label = NULL,
  value = NULL,
  placeholder = NULL
)
}
\arguments{
\item{session}{The \verb{session`` object passed to function given to  }shinyServer\verb{. Default is }getDefaultReactiveDomain()`.}

\item{input_id}{The id of the input object.}

\item{label}{The label to set for the input object.}

\item{value}{Initial value.}

\item{placeholder}{A character string giving the user a hint as to
what can be entenred into the control.}
}
\description{
Change the value of a SunEditor input on the client
}
