% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_sun_editor.R
\name{use_sun_editor}
\alias{use_sun_editor}
\title{Use Sun Editor}
\usage{
use_sun_editor()
}
\description{
This function imports the necessary JS and CSS files for the Sun Editor.
It uses the singleton function to ensure that the files are only imported once.
The function does not take any parameters and does not return any value.
}
